Move DSN helpers to its own package: database/dsn
