+++ database/dsn/dsn_test.go
@@ -1,4 +1,4 @@
-package helpers
+package dsn
 
 import (
 	"net/url"
