+++ migration/migrator/migrator_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"testing"
 
-	"github.com/cloudquery/cq-provider-sdk/helpers"
 	"github.com/cloudquery/cq-provider-sdk/provider/schema"
 
 	"github.com/golang-migrate/migrate/v4"
@@ -160,13 +159,4 @@ func TestFindLatestMigration(t *testing.T) {
 	mv, err = mtest.FindLatestMigration("latest")
 	assert.Nil(t, err)
 	assert.Equal(t, uint(5), mv)
-}
-
-func TestParseConnectionString(t *testing.T) {
-	url, err := helpers.ParseConnectionString("postgres://postgres:pass@localhost:5432/postgres?sslmode=disable")
-	assert.Nil(t, err)
-	assert.Equal(t, "postgres://postgres:pass@localhost:5432/postgres?sslmode=disable", url.String())
-	url, err = helpers.ParseConnectionString("host=localhost user=postgres password=pass database=postgres port=5432 sslmode=disable")
-	assert.Nil(t, err)
-	assert.Equal(t, "postgres://postgres:pass@localhost:5432/postgres?sslmode=disable", url.String())
 }
